Move plugin Implementation to sfdc_input_plugin.rb from sfdc.rb
Because filename should be matched class name in which implemented, but
Embulk load `plugin-name.rb`.
